Docs improvements
I would also consider updating the name of the `options` prop to `params` in the future (or allow configuration of the `Socket` itself). I found it a bit confusing and wasn't sure whether that prop configured the `Socket` itself, or was just the params. Turns out it's just the params:

```javascript
const socket = new Socket(wsUrl, { params: options })
```